+++ src/main/java/com/fuseinterns/libraryManagementSystem/user/UserController.java
@@ -18,10 +18,14 @@ public class UserController {
 	
 	@RequestMapping(value="/api/user",method=RequestMethod.GET, produces=MediaType.APPLICATION_JSON_VALUE)
 //note responseEntity is used to provide status 	
-	public ResponseEntity<List<User>> getAllUser()
+	public ResponseEntity<List<User>> getAllUser(@RequestHeader(value = "userId")String userId, @RequestHeader(value = "password")String password)
 	{
-		//note as ResponseEntity is used so new is a must 
-		return new ResponseEntity<>(this.userService.getAllUser(), HttpStatus.OK);
+		//note as ResponseEntity is used so new is a must
+		User currentUser = userService.getUserById(userId);
+		if(currentUser!=null && currentUser.getPassword().matches(password) && currentUser.getRole().toLowerCase().equals("admin")) {
+			return new ResponseEntity<>(this.userService.getAllUser(), HttpStatus.OK);		}
+			return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+
 	}
 	
 	@RequestMapping(value="/api/user",method=RequestMethod.POST, produces=MediaType.APPLICATION_JSON_VALUE )
@@ -35,13 +39,20 @@ public class UserController {
 	}
 	
 	@RequestMapping(value = "/api/user/{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public  void deleteUser(@PathVariable String id){
-         this.userService.deleteUser(id);
+    public  void deleteUser(@RequestHeader(value = "userId")String userId, @RequestHeader(value = "password")String password,@PathVariable String id){
+		User currentUser = userService.getUserById(userId);
+		if(currentUser!=null && currentUser.getPassword().matches(password) && currentUser.getRole().toLowerCase().equals("admin")) {
+			this.userService.deleteUser(id);
+		}
     }
 	
 	 @RequestMapping(value = "/api/user/{id}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
-	    public ResponseEntity<User> getUserById(@PathVariable String id){
-	        return new ResponseEntity<>(this.userService.getUserById(id), HttpStatus.OK);
+	    public ResponseEntity<User> getUserById(@RequestHeader(value = "userId")String userId, @RequestHeader(value = "password")String password,@PathVariable String id){
+		 User currentUser = userService.getUserById(userId);
+		 if(currentUser!=null && currentUser.getPassword().matches(password) && currentUser.getRole().toLowerCase().equals("admin")) {
+			 return new ResponseEntity<>(this.userService.getUserById(id), HttpStatus.OK);
+		 }
+		 return new ResponseEntity<>(HttpStatus.FORBIDDEN);
 	    }
 
 }
